[BE] Move indexing tests to test_indexing
Which enables them on MPS device

[ghstack-poisoned]
